+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-v 0.10
+v 0.10.1
   - Use built-in interactive shell for openbsd and darwin builds. #50
+
+v 0.10
   - Temporary do not build for openbsd and darwin targets. #48
   - Do not write empty buffer to file. #44
   - Improve the interactive prompt. #43